part_02

+++ part_01/Exercise_11.java
@@ -17,14 +17,15 @@ public class Exercise_11 {
 
     public static void main(String[] args) {
 
-        double currentPopulation = 380123456;
-        double born = 31543200 / 6;
-        double deaths = 31543200 / 12;
-        double immigrates = 31543200 / 40;
+        int currentPopulation = 380123456;
+        // 31543200 is total seconds in a year
+        int born = 31543200 / 6;
+        int deaths = 31543200 / 12;
+        int immigrates = 31543200 / 40;
 
-        double bornB = born * 3;
-        double deathsD = deaths * 3;
-        double immigratesI = immigrates * 3;
+        int bornB = born * 3;
+        int deathsD = deaths * 3;
+        int immigratesI = immigrates * 3;
 
         System.out.println("Total population for the next three years: " + (currentPopulation - deathsD + (bornB + immigratesI)));
